Move storageImageDestination into a separate storage/storage_dest.go
storage_image.go is becoming too large, and we'd like to
have easy access to both internal/imagesource/impl and
internal/imagedestination/impl, which is easier with separate files.

Only moves unchanged code, should not change behavior.
